pyproject.toml
[tool.poetry]
name = "invest"
version = "0.1.0"
description = "Invest App"
authors = ["sandropark <birdob@gmail.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
pyyaml = "^6.0.2"
requests = "^2.32.3"
pytz = "^2024.2"
finance-datareader = "^0.9.94"
pandas-datareader = "^0.10.0"
bs4 = "^0.0.2"
plotly = "^5.24.1"
yfinance = "^0.2.49"
schedule = "^1.2.2"
apscheduler = "^3.10.4"
sqlalchemy = "^2.0.36"
pyupbit = "^0.2.34"
ipykernel = "^6.29.5"
black = "^24.10.0"
exchange-calendars = "^4.6"
dependency-injector = "^4.44.0"
pymysql = "^1.1.1"
groq = "^0.14.0"
alembic = "^1.14.0"
ccxt = "^4.4.52"
gspread = "^6.2.0"
gspread-dataframe = "^4.0.0"
pandas-stubs = "^2.2.3.250308"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = [".", "src"]

[tool.black]
line-length = 150